requirements.txt
pystray==0.19.5
Pillow==10.3.0
PyYAML==6.0.1
requests==2.31.0
PyInstaller==6.6.0
